Porting of SQL-related functions
