+++ module/ldbc-dsl/src/main/scala/ldbc/dsl/logging/LogHandler.scala
@@ -4,7 +4,7 @@
  * For more information see LICENSE or https://opensource.org/licenses/MIT
  */
 
-package ldbc.sql.logging
+package ldbc.dsl.logging
 
 import cats.Applicative
